Recreating resource storage.....
This allows to use a lot less resources. The system will use base files and apply color modifiers during rendering. Created all the storage systems for this system.

Initializer has to be changed still.
Files have to be renamed and other deleted.
Textures may need an lighter overlay (Maybe in the color direction of Iron).