Preparing for initial testing....
Finished the rewrite. Now made a test body for the preparegame function inside the initializer. This allows testing of the basic rendering mechanics. Need to take a look into the crafting system (aka if there is an event for crafting), cause i will use the vanilla crafting bench on initial release.

A better crafting system will be made in the future.