Started working on an ANVIL Gui and switch Component based UI Elements. Core of the Component system is done. Still needs some bugfixing.
Ported the current Ledger system to the Component one.

Components still to be made:
Slot
Burnflame
Progressarrow
Fluidtank
Textbox
Backgroundborder

Extracting some rendering work out of the initial Component classes into a Helper class, as part of the client util package. Still working on the UI for the Anvil, in specific the UI Environment Awareness.