First pass on the BlackSmithsAnvil
Finished the VanillaCrafting mechanics of the BlackSmithsAnvil, could still have improved mechanics but will work for now.
Finished the basic GUI of the BlackSmithsAnvil.
Updated all TileEntities and Slots to use the new IItemStorage system instead of Vanilla's IInventory.
Made the HeatableItemRegistry store a list of HeatableItemTypes on the reverse Fluidstack map, instead of outputting a warning to the log and skipping.
Fixed the BlackSmithsAnvil logo not being logo into